Separated house building into its own file
